requirements-gui.txt
# Desktop Use Engine (src/desktop_use/) — Windows only
# These packages require native dependencies that don't build in Docker Linux containers
mss>=9.0.0
pyautogui>=0.9.54
winocr>=0.2.0               # system-native OCR (WinRT)
Pillow>=10.0.0
rapidfuzz>=3.0.0
pynput>=1.7.6
pywinauto>=0.6.8
